test-automation/requirements.txt
httpx>=0.24.0
rich>=13.0.0
pyyaml>=6.0
click>=8.1.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0  # For Gemini AI analysis
a2a>=0.1.0  # Official A2A Protocol SDK